+++ erp_backend/priv/repo/migrations/20210224035415_create_orders.exs
@@ -4,7 +4,7 @@ defmodule Erp.Repo.Migrations.CreateOrders do
   def change do
     create table(:orders) do
       add :orderId, :id
-      add :price, :integer
+      add :price, :float
       add :userEmail, references(:users, column: :email, type: :string, on_delete: :nothing)
       add :time, :naive_datetime
       add :bikesAmount, :integer
